dag-generator/containers/load_bigquery/requirements.txt
google-cloud-storage>=1.23.0
google-cloud-bigquery>=1.22.0